// src/lib/hive/types.ts
export interface HiveComment {
  id: number;
  author: string;
  permlink: string;
  category: string;
  parent_author: string;
  parent_permlink: string;
  title: string;
  body: string;
  json_metadata: string;
  last_update: string;
  created: string;
  active: string;
  last_payout: string;
  depth: number;
  children: number;
  net_rshares: string;
  abs_rshares: string;
  vote_rshares: string;
  children_abs_rshares: string;
  cashout_time: string;
  max_cashout_time: string;
  total_vote_weight: string;
  reward_weight: number;
  total_payout_value: string;
  curator_payout_value: string;
  author_rewards: string;
  net_votes: number;
  root_author: string;
  root_permlink: string;
  max_accepted_payout: string;
  percent_hbd: number;
  allow_replies: boolean;
  allow_votes: boolean;
  allow_curation_rewards: boolean;
  url: string;
  root_title: string;
  pending_payout_value: string;
  total_pending_payout_value: string;
  active_votes: Array<{
    voter: string;
    weight: number;
    rshares: string;
    percent: number;
    reputation: string;
    time: string;
  }>;
  replies: string[];
  author_reputation: string;
  promoted: string;
  body_length: number;
  reblogged_by: string[];
}